rag_texto/requirements.txt
python-dotenv
langchain
langchain-community
langchain-chroma
chromadb
pypdf
